Use PCM.
